requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
structlog==23.2.0
pyyaml==6.0.1
asyncio-mqtt==0.16.1

# LLM and AI
openai==1.3.7
google-generativeai==0.3.2
sentence-transformers==2.2.2
huggingface-hub==0.20.3
alembic==1.13.1
redis==5.0.1
aiosqlite==0.19.0

# Enhanced Agent Dependencies
numpy==1.24.3
aiohttp==3.9.1
redis[async]==5.0.1
dataclasses-json==0.6.3

# Logging and Monitoring
prometheus-client==0.19.0
grafana-api==1.0.3

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Security
keyring==24.3.0
cryptography==41.0.7
google-auth==2.23.4
cryptography==41.0.8
bcrypt==4.1.2

# Performance and Caching
aioredis==2.0.1
aiomcache==0.7.0

# Database and Storage
sqlalchemy==2.0.23
alembic==1.13.1
aiosqlite==0.19.0

# Vector Database
chromadb==0.4.18

# Web Framework Extensions
fastapi-cors==0.0.6
fastapi-limiter==0.1.5

# Configuration and Environment
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Async Utilities
aiofiles==23.2.1
asyncio-throttle==1.0.2

# HTTP Client
httpx==0.25.2

# Data Processing
pandas==2.1.4
numpy==1.24.3

# Machine Learning
scikit-learn==1.3.2
scipy==1.11.4

# Monitoring and Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-exporter-prometheus==1.21.0

# Testing and Quality
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-benchmark==4.0.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Development Tools
pre-commit==3.6.0
isort==5.12.0

# Tool Registry Dependencies
jsonschema==4.20.0
orjson==3.9.10 